[PMSM_FOC_ROLO_1shunt] Regenerated pmsm_foc_rolo_1shunt_sam_c21_mclv2 application
MH3-36548

+++ apps/pmsm_foc_rolo_1shunt/firmware/src/config/mclv2_sam_c21_pim/peripheral/adc/plib_adc1.c
@@ -99,7 +99,7 @@ void ADC1_Initialize( void )
     ADC1_REGS->ADC_REFCTRL = ADC_REFCTRL_REFSEL_INTVCC2;
 
     /* Input pin */
-    ADC1_REGS->ADC_INPUTCTRL = ADC_POSINPUT_AIN10;
+    ADC1_REGS->ADC_INPUTCTRL = (uint16_t) ADC_POSINPUT_AIN10;
 
     /* Resolution & Operation Mode */
     ADC1_REGS->ADC_CTRLC = ADC_CTRLC_RESSEL_12BIT | ADC_CTRLC_WINMODE(0) ;
@@ -142,7 +142,7 @@ void ADC1_Disable( void )
 void ADC1_ChannelSelect( ADC_POSINPUT positiveInput, ADC_NEGINPUT negativeInput )
 {
     /* Configure pin scan mode and positive and negative input pins */
-    ADC1_REGS->ADC_INPUTCTRL = positiveInput | negativeInput;
+    ADC1_REGS->ADC_INPUTCTRL = (uint16_t) positiveInput | (uint16_t) negativeInput;
 
     while((ADC1_REGS->ADC_SYNCBUSY & ADC_SYNCBUSY_INPUTCTRL_Msk) == ADC_SYNCBUSY_INPUTCTRL_Msk)
     {
